# src/main/java/com/blokgpt/model/CourtsBO.java
package com.blokgpt.model;

import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class CourtsBO {

	private int courtId;
	private String courtName;
	private String addressLine1;
	private String AddressLine2;
	private String city;
	private String postalCode;
	private String state;
	
	
}