# src/main/java/com/blokgpt/model/CustomersBO.java
package com.blokgpt.model;

import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class CustomersBO {

	
	private int customerId;
	private String customerName;
	private String addressLine1;
	private String addressLine2;
	private String city;
	private String mobileNumber;
	private String postalcode;
	private String state;
}